+++ db/migrate/20161116222637_rename_options_to_deprecated_options_for_domains.rb
@@ -1,0 +1,5 @@
+class RenameOptionsToDeprecatedOptionsForDomains < ActiveRecord::Migration[5.0]
+  def change
+    rename_column :domains, :options, :deprecated_options
+  end
+end