+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Check results are now precomputed to improve the speed of the sheets index
 - **Design Changes**
   - Improved filtering designs by name on design index
+- **Domain Changes**
+  - Improved how domain options are stored and updated
 - **Sheet Changes**
   - Improved display of changes to signature variables in sheet transactions
   - Question styling now alternates dynamically based on hidden questions